requirements.txt
mlflow
python-dotenv
joblib
pandas
scikit-learn==1.3.2
requests
dagshub
apache-airflow